LayoutTests/http/tests/security/contentSecurityPolicy/object-src-none-blocked-expected.txt

CONSOLE MESSAGE: Refused to load the object 'data:application/x-webkit-test-netscape,alertwhenloaded' because it violates the following Content Security Policy directive: "object-src 'none'".



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
